modification accoding to Claude review
