+++ sentence-embedding/python-lib/dku_language_model/__init__.py
@@ -1,2 +1,2 @@
-from dku_language_model.context_independent_language_model import FasttextModel, Word2vecModel, GloveModel
-from dku_language_model.contextual_language_model import ElmoModel
+from context_independent_language_model import FasttextModel, Word2vecModel, GloveModel
+from contextual_language_model import ElmoModel
