+++ sentence-embedding/python-lib/dku_language_model/context_independent_language_model.py
@@ -4,6 +4,7 @@ from collections import Counter
 from gensim.models import KeyedVectors
 from sklearn.decomposition import TruncatedSVD
 from abstract_language_model import AbstractLanguageModel
+from language_model_utils import clean_text
 import logging
 logger = logging.getLogger(__name__)
 
@@ -25,9 +26,9 @@ class ContextIndependentLanguageModel(AbstractLanguageModel):
         return avg_embedding
         
     def get_sentence_embedding(self, texts):
-        cleaned_texts = map(self.clean_text, texts)
-        res = map(self.compute_average_embedding, cleaned_texts)
-        return res
+        cleaned_texts = map(clean_text, texts)
+        embeddings = map(self.compute_average_embedding, cleaned_texts)
+        return embeddings
         
     def get_weighted_sentence_word_vectors(self, text, weights):
         #Check if sentence contains at least one token and return None if not
@@ -42,7 +43,7 @@ class ContextIndependentLanguageModel(AbstractLanguageModel):
         avg_embedding = np.mean(embeddings, axis=0)
         return avg_embedding
     
-    def remove_first_principal_component(self, X, npc):
+    def remove_principal_components(self, X, npc):
         """Removes the first PC for computing SIF."""
         svd = TruncatedSVD(n_components=npc, n_iter=7, random_state=0)
         X = np.array(X)
@@ -50,19 +51,19 @@ class ContextIndependentLanguageModel(AbstractLanguageModel):
         u = svd.components_
         return X - X.dot(u.T).dot(u)
 
-    def contruct_final_res(self, res,is_void):
-        res_final = []
+    def contruct_final_embeddings(self, embeddings, is_void):
+        final_embeddings = []
         j = 0
         for v in is_void:
             if v == 0:
-                res_final.append(res[j])
+                final_embeddings.append(embeddings[j])
                 j+=1
             else:
-                res_final.append(np.nan)
-        return res_final
+                final_embeddings.append(np.nan)
+        return final_embeddings
 
     def get_weighted_sentence_embedding(self, texts, smoothing_parameter, npc):
-        cleaned_texts = map(self.clean_text, texts)
+        cleaned_texts = map(clean_text, texts)
 
         # Compute word weights
         word_weights = Counter()
@@ -72,21 +73,22 @@ class ContextIndependentLanguageModel(AbstractLanguageModel):
         n_words = float(sum(word_weights.values()))
         for k, v in word_weights.items():
             word_weights[k] /= n_words
-            word_weights[k] = float(smoothing_parameter) / (smoothing_parameter + word_weights[k])
+            word_weights[k] = smoothing_parameter / (smoothing_parameter + word_weights[k])
 
         # Compute SIF
         logger.info("Computing weighted average embeddings...")
-        res = map(lambda s: self.compute_weighted_average_embedding(s, word_weights), cleaned_texts)
+        raw_embeddings = np.array(map(lambda s: self.compute_weighted_average_embedding(s, word_weights), cleaned_texts))
 
         # Remove empty sentences and save their indecies
-        is_void = map(lambda x: 1 if x.shape == () else 0 , res)
-        res = [x for x,y in zip(res,is_void) if y==0]
+        is_void = np.array(map(lambda x: x.shape == (), raw_embeddings))
+        embeddings_without_void = [x for x,y in zip(raw_embeddings,is_void) if y==0]
 
-        logger.info("Removing vectors first principal component...")
-        res = self.remove_first_principal_component(res, npc)
-        res = self.contruct_final_res(res,is_void)
+        logger.info("Removing vectors principal component...")
+        embeddings_processed = self.remove_principal_components(embeddings_without_void, npc)
+        #TODO: refactor this to avoid the weird j index condition 
+        final_embeddings = self.contruct_final_embeddings(embeddings_processed, is_void)
             
-        return res
+        return final_embeddings
     
     
 class Word2vecModel(ContextIndependentLanguageModel):
@@ -98,6 +100,7 @@ class Word2vecModel(ContextIndependentLanguageModel):
     def load_model(self):
         logger.info('Loading Word2Vec model...')
         model = KeyedVectors.load_word2vec_format(self.model_path, binary=True)
+        logger.info('Done')
         self.word2idx = {w: i for i, w in enumerate(model.index2word)}
         self.embedding_matrix = model.vectors
 
@@ -124,6 +127,7 @@ class FasttextModel(ContextIndependentLanguageModel):
 
                 embedding = np.array(vector).astype(float)
                 embedding_matrix.append(embedding)
+        logger.info('Done')
         self.embedding_matrix = np.array(embedding_matrix)
         self.word2idx = word2idx
