+++ sentence-embedding/python-runnables/download-pretrained-embedding/runnable.py
@@ -27,7 +27,6 @@ def get_confirm_token(response):
     for key, value in response.cookies.items():
         if key.startswith('download_warning'):
             return value
-
     return None
